// lib/liveAgent.types.ts
export interface components {
    schemas: {
// ...
        DomainIn: {
            /** Id */
            id: string;
            /**
             * Interviewercontext
             * @default
             */
            interviewerContext?: string;
            /**
             * Scoringcontext
             * @default
             */
            scoringContext?: string;
            /**
             * Roundhints
             * @default
             */
            roundHints?: string;
            /**
             * Roadmaphints
             * @default
             */
            roadmapHints?: string;
            /**
             * Version
             * @default 1
             */
            version?: number;
        };
// ...
        PersonaIn: {
            /** Name */
            name: string;
            /** Tone */
            tone: string;
        };
// ...
    };
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}